+++ src/game/nlcore/elements/story.ts
@@ -10,8 +10,8 @@ import {Storable} from "@core/elements/persistent/storable";
 
 /* eslint-disable @typescript-eslint/no-empty-object-type */
 export type StoryConfig = {};
+/**@internal */
 export type ElementStateRaw = Record<string, any>;
-export type NodeChildIdMap = Map<string, string>;
 
 export class Story extends Constructable<
     SceneAction<"scene:action">,
@@ -21,6 +21,11 @@ export class Story extends Constructable<
     static defaultConfig: StoryConfig = {};
     /**@internal */
     static MAX_DEPTH = 10000;
+
+    /**@internal */
+    public static empty(): Story {
+        return new Story("empty").entry(new Scene("empty"));
+    }
 
     /**@internal */
     readonly name: string;