+++ src/game/player/libElements.ts
@@ -1,7 +1,8 @@
 import Isolated from "@player/lib/isolated";
 import Say from "@player/elements/say/Say";
-import {Top, Center, Bottom} from "@player/lib/PlayerFrames";
+import {Top, Center, Bottom, Full} from "@player/lib/PlayerFrames";
 import {VBox, HBox} from "@player/lib/UtilComponents";
+import {usePreference} from "@player/lib/preferences";
 
 export {
     Isolated,
@@ -11,4 +12,6 @@ export {
     Bottom,
     VBox,
     HBox,
+    Full,
+    usePreference,
 };